+++ src/backend/data/models.json
@@ -1,0 +1,24 @@
+{
+    "models": [
+      {
+        "id": 1,
+        "name": "Module"
+      },
+      {
+        "id": 2,
+        "name": "Phase"
+      },
+      {
+        "id": 3,
+        "name": "Task"
+      },
+      {
+        "id": 4,
+        "name": "Step"
+      },
+      {
+        "id": 5,
+        "name": "Guidance"
+      }
+    ]
+}
