+++ src/backend/routes/index.js
@@ -1,0 +1,11 @@
+const express = require('express');
+
+const router = express.Router();
+
+router.use('/api/modules', require('./modules'));
+router.use('/api/phases/:moduleId', require('./phases'));
+router.use('/api/tasks/:phaseId', require('./tasks'));
+
+// @todo error handlings
+
+module.exports = router;
